feat(matching): delete legacy matchingEngine files and add lint guard - TER-1250
- Delete server/matchingEngine.ts (593 lines, no remaining consumers)
- Delete server/matchingEngineReverseSimplified.ts (325 lines)
- Inline findClientNeedsForBatch and findClientNeedsForVendorSupply into
  matchingEngineEnhanced.ts (the canonical engine)
- Add clients/vendors to schema imports in matchingEngineEnhanced.ts
- Add no-restricted-imports ESLint rule blocking both deleted filenames
  to prevent reintroduction

All consumers already imported from matchingEngineEnhanced (TER-1249/PR #656).
This is the final cleanup step for TER-1195 matching engine consolidation.
