+++ docs/sessions/active/TER-1250-session.md
@@ -1,5 +1,36 @@
 # TER-1250 Agent Session
 - Ticket: TER-1250
 - **Branch:** `feat/ter-1250-delete-legacy-matching-engine`
-- Status: In Progress
-- Agent: Factory Droid
+- Status: Complete
+- Agent: Manus PM
+
+## Summary
+TER-1250 requested deleting `server/matchingEngine.ts` and `server/matchingEngineReverseSimplified.ts`
+and adding a forbidden-import lint rule to prevent them from reappearing.
+
+## Work Completed
+1. ✅ Deleted `server/matchingEngine.ts` (593 lines, no remaining consumers)
+2. ✅ Deleted `server/matchingEngineReverseSimplified.ts` (325 lines)
+   - Inlined `findClientNeedsForBatch` and `findClientNeedsForVendorSupply` into `matchingEngineEnhanced.ts`
+   - Added `clients` and `vendors` to schema imports in `matchingEngineEnhanced.ts`
+3. ✅ Added `no-restricted-imports` ESLint rule in `eslint.config.js` blocking both legacy filenames
+4. ✅ All consumers already use `matchingEngineEnhanced` (verified by grep)
+
+## Changes
+- `server/matchingEngineEnhanced.ts`: Added `clients`/`vendors` imports + inlined reverse matching functions
+- `server/matchingEngine.ts`: DELETED
+- `server/matchingEngineReverseSimplified.ts`: DELETED
+- `eslint.config.js`: Added `no-restricted-imports` guard for both deleted files
+- `docs/sessions/active/TER-1250-session.md`: Session documentation
+
+## Acceptance Criteria Status
+- ✅ server/matchingEngine.ts is deleted
+- ✅ server/matchingEngineReverseSimplified.ts is deleted
+- ✅ No remaining imports of either legacy file anywhere in the codebase
+- ✅ ESLint no-restricted-imports rule blocks these filenames from reappearing
+- ✅ PR opened against main (PR #660)
+
+## References
+- Parent ticket: TER-1195 (matching engine consolidation)
+- Blocked by: TER-1249 (consumer migration) — MERGED as PR #656
+- This PR: #660
